Updated default route, layout, styles, and added AccountController
Updated the default controller route pattern to direct to the AccountController by default. Made several changes to the layout, including adding a new stylesheet for bootstrap icons, updating the header, and adding a new offcanvas component for side navigation. Updated CSS for body and header, and added media queries for smaller screens. Added a new tooltip class in site.css. Introduced a new AccountController with Index, Login, and Register actions, and corresponding views. Added a new PNG image file for use in the application's menu.