Update Register form and related CSS, modify layout file
Significant modifications have been made to the `Register.cshtml` file, including a change in the model used for the view to `UserModel`, a redesign of the form layout using flexbox, addition of new form fields, and inclusion of validation and bilingual labels. The `Register.cshtml.css` file has been updated with new styles and media queries. The `_Layout.cshtml` file has seen a change in the Swahili translation of the app name. The `site.css` file has been updated with a new background color, updated font-face declaration for 'Fira Code VF', and removal of styles for `.tooltip` and `.tooltip .tooltiptext`.