Updated views, controllers, and scripts for better UX
In this commit, several changes were made to improve the user experience. A new action method `Results` was added to `TestController.cs` to display test results. The `AddVenue` method in `VenuesController.cs` was modified to return a specific view.

In `Index.cshtml`, a modal dialog was removed and the action of the "Create venue" button was updated. The title of `EnglishQuestion.cshtml` was changed for clarity.

In `Venue.cshtml`, the id of a div was updated and reflected in the associated CSS file. A new hidden element was added to `timer.js` to be shown when the timer expires.

Finally, a new view `Results.cshtml` and its associated CSS file were added to display test results.