Updated session details and user list fetching
Updated the `Details` method in `SessionsController.cs` to be asynchronous and accept additional parameters. Added a call to `apiClient.GetSessionUsers` to fetch a list of users for a session. Implemented a new method `GetSessionUsers` in `ApiClient.cs` to make an asynchronous GET request to fetch users. Updated `Details.cshtml` to display additional session details and a static list of users. Added route parameters to the link in `Venue.cshtml`. Added a new record `UserData` in `RSAllies.Contracts.Responses`. Added new CSS styles for the `Details` view.