`Minor change in SessionsController.cs parameters`
` `

`Made a minor change in the parameters of the `Details` method in the `SessionsController` class. The change does not affect the functionality of the code, as the removed line is identical to the added line. The parameters remain as `string id`, `string startTime`, `string endTime`, `string date`, `string capacity`, `string venueName`. This appears to be a redundant change.`