Add .NET Core WebApp deployment workflow
This commit introduces a new GitHub Actions workflow for deploying .NET Core applications to Azure Web Apps, named "Build and deploy .NET Core application to Web App Dctt". It triggers on pushes to the `master` branch, contrasting with the previous workflow which targeted the `main` branch for Azure Container Apps deployment. Key enhancements include the integration of a testing phase during the build job, the adoption of the `azure/webapps-deploy@v2` action for deployment using a publish profile, and updates to the setup-dotnet action version from `v1.8.0` to `v3`. Additionally, this workflow specifies a working directory and a package path, aiming for a more organized project structure. These changes signify a shift in deployment targets from Azure Container Apps to Azure Web Apps and introduce a more structured and test-inclusive CI/CD process.